[feat] 입력 받은 숫자가 Integer.MAX_VALUE보다 작은지 판단하는 메서드 구현

+++ src/main/java/calculator/util/OutputUtil.java
@@ -9,7 +9,7 @@ public class OutputUtil {
         System.out.print("결과 : ");
     }
 
-    public void printResult(int result){
+    public void printResult(Long result){
         System.out.print(result);
     }
